s04e05/requirements.txt
python-dotenv
requests
deepagents
langchain
langchain-openai
